Refactor how Lab's are handled
